added empty screen for a new trip

+++ app/src/main/java/com/newvisadatescalculator/routes/TripListRoute.kt
@@ -9,7 +9,7 @@ import com.visadatescalculator.viewmodel.DaysCalculatorViewModel
 
 @Composable
 fun TripListRoute(
-    onNavigateToAddTrip: () -> Unit,
+    onNavigateToAddTrip: (Int) -> Unit,
     viewModel: DaysCalculatorViewModel = hiltViewModel()
 ) {
